Ensure a list of same labels can be associated with multiple variables
This commit remove a call to "REMOVE_DUPLICATES" that was probably added
only "to be safe". It turns out that this call is (1) not needed and
(2) it also had the undesired side effect of removing the association
of a variable with additional labels.

For example, if two variable FOO and BAR where associated
with labels "LabelA" and "LabelB":

project(MyProj)
[...]
set(BAR "ThisIsBar")
set(FOO "ThisIsFoo")
mark_as_superbuild(VARS FOO LABELS LabelA LabelB)
mark_as_superbuild(VARS BAR LABELS LabelA LabelB)

The corresponding arguments list generated in '_sb_get_external_project_arguments'
was:

-DFOO:STRING=ThisIsFoo
-DBAR:STRING=ThisIsBar
-DMyProj_EP_LABEL_LabelA:STRING=FOO;BAR
-DMyProj_EP_LABEL_LabelB:STRING=FOO;BAR

Then, the call to "list(REMOVE_DUPLICATES _all)" was reducing this list
to:

-DFOO:STRING=ThisIsFoo
-DBAR:STRING=ThisIsBar
-DMyProj_EP_LABEL_LabelA:STRING=FOO;BAR
-DMyProj_EP_LABEL_LabelB:STRING=FOO             <----- BAR is expected :(


.. which was the source of our problem.